Train/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen3omni-ms-swift-train"
version = "0.1.0"
description = "Minimal modular ms-swift training launcher for Qwen3-Omni."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
]

[project.scripts]
q3o-train = "qwen3omni_train.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
